+++ src/browser.ts
@@ -2,3 +2,4 @@
 export * from './version';
 export * from './random';
 export * from './drawbox';
+export * from './combination';